catch the orphaned nodes
It can happen that Tetrad learns a forbidden edge. Argh. To catch this, we delete forbidden edges from any insertion. But then it can happen that a node has no edge at all, not even with an empty parent. In that case the Bif generator gets messed up. So we catch such orphaned nodes in the next statement.